smaller-labse-tf-serving/requirements.txt
tensorflow<2.8
tensorflow-text<2.8
absl-py
